+++ src/jupyter_starters/types.py
@@ -1,3 +1,13 @@
-""" some types and contstants
+""" some types and constants
 """
+# pylint: disable=too-few-public-methods
 NS = "starters"
+
+
+class Status:
+    """ pseudo-enum for managing statuses
+    """
+
+    DONE = "done"
+    CONTINUING = "continuing"
+    ERROR = "error"